# api/swagger.yaml
openapi: 3.0.3
info:
  title: BREECE API SPECIFICATION
  description: |-
    This is the documentation made by Akbar Fikri. You can find all about my API Specification for BREECE Apps.
    If you have criticism or suggestions, don't hesitate to let me know via Github issue

    Some useful links:
    - [Github Repository](https://github.com/AkbarFikri/BREECE-BE)
    - [BREECE Website](breece.ub.ac.id)
  # termsOfService: http://swagger.io/terms/
  contact:
    email: akbarfikriabdillah@gmail.com
  license:
    name: MIT License
    url: https://opensource.org/license/mit
  version: 1.0.0
...
servers:
  - url: https://breece.akbarfikri.site/api/v1
...
paths:
...
  /event:
    get:
      tags:
        - Event
      summary: its use for fetch event data.
      operationId: fetchEvent
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema:
            type: string
            example: Orasi
          description: Its use for search feature
        - in: query
          name: place
          schema:
            type: string
            example: NKRI
          description: Its use for filter by place. It can be dynamic by input user
        - in: query
          name: date
          schema:
            type: string
            example: 2022-10-3
            format: YY-MM-DD
          description: Its use for filter by date.
        - in: query
          name: category
          schema:
            type: string
            example: asdwadlhaodiaw-dashdowadasdlk-asdjawduh
            format: uuid
          description: Its use for filter by category. Use the category_id that provided from category endpoint
        - in: query
          name: page
          schema:
            type: number
            example: 1
          description: Its use for pagination, per page is contain 10 object. Default value is 1.

      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Successfully to find all events
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        title:
                          type: string
                        description:
                          type: string
                        speakers:
                          type: array
                          items: {}
                        speaker_roles:
                          type: array
                          items: {}
                        place:
                          type: string
                        date:
                          type: string
                          format: date-time
                        start_at:
                          type: string
                          format: date-time
                        organize_by:
                          type: string
                          format: uuid

        "404":
          description: Success but data record is not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Event with data provided is not found.
                  data:
                    type: array
                    example: []
        "500":
          description: Internal server error
    post:
      tags:
        - Event
      summary: Its use for post new event. Provided just for organizer
      operationId: postNewEvent
      security:
        - bearerAuth: []
      requestBody:
        description: Use the ID from user
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - "title"
                - "description"
                - "speakers"
                - "speaker_roles"
                - "date"
                - "start_at"
                - "category_id"
                - "place"
                - "is_public"
              properties:
                banner:
                  type: string
                  format: binary
                title:
                  type: string
                description:
                  type: string
                speakers:
                  type: array
                speaker_roles:
                  type: array
                date:
                  type: string
                  format: date
                start_at:
                  type: string
                  format: date-time
                category_id:
                  type: string
                  format: uuid
                link:
                  type: string
                place:
                  type: string
                is_public:
                  type: boolean

        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Successfully to verified profile
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      title:
                        type: string
                      description:
                        type: string
                      speakers:
                        type: array
                        items: {}
                      speaker_roles:
                        type: array
                        items: {}
                      place:
                        type: string
                      date:
                        type: string
                        format: date-time
                      start_at:
                        type: string
                        format: date-time
                      organize_by:
                        type: string
                        format: uuid
        "400":
          description: Invalid Request data
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Bad request, data provided is invalid
                  data:
                    type: object
                    example:
        "500":
          description: Internal server error
  /event/{id}:
    get:
      tags:
        - Event
      summary: Its use for fetch event details
      operationId: getEventDetails
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Successfully to verified profile
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      title:
                        type: string
                      description:
                        type: string
                      speakers:
                        type: array
                        items: {}
                      speaker_roles:
                        type: array
                        items: {}
                      place:
                        type: string
                      date:
                        type: string
                        format: date-time
                      start_at:
                        type: string
                        format: date-time
                      organize_by:
                        type: string
                        format: uuid
        "400":
          description: Invalid Request data
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Bad request, data provided is invalid
                  data:
                    type: object
                    example:
        "500":
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT